fixes for Linux build (gcc 4.4.4) - max macro is undefined, use MAX macro instead - define x86_64 as LITTLE_ENDIAN - va_list variables should be reset (va_end && va_start) after using, otherwise I have segfault in StringX.cpp:Format - different types (size_t != unsigned int) in declaration and implementation pws_os::asctime - const function modifiers - slackBuild script added
git-svn-id: svn://svn.code.sf.net/p/passwordsafe/code/trunk@3532 1f79f812-37fb-46fe-a122-30589dd2bf55

+++ pwsafe/pwsafe/src/os/linux/pws_time.cpp
@@ -18,7 +18,7 @@ struct tm *gmtime64_r(const __time64_t *timep, struct tm *result)
   return gmtime_r((const time_t *)timep, result);
 }
 
-int pws_os::asctime(TCHAR *s, unsigned int, tm const *t)
+int pws_os::asctime(TCHAR *s, size_t, tm const *t)
 {
 #ifdef UNICODE
   char cbuf[26]; // length specified in man (3) asctime
